+++ fpakman/core/system.py
@@ -1,5 +1,8 @@
+import os
 import subprocess
 from typing import List
+
+from fpakman.core import resource
 
 
 def run_cmd(cmd: str, expected_code: int = 0, ignore_return_code: bool = False) -> str:
@@ -9,3 +12,8 @@ def run_cmd(cmd: str, expected_code: int = 0, ignore_return_code: bool = False) 
 
 def stream_cmd(cmd: List[str]):
     return subprocess.Popen(cmd, stdout=subprocess.PIPE, env={'LANG': 'en'}).stdout
+
+
+def notify_user(msg: str, icon_path: str = resource.get_path('img/flathub_45.svg')):
+    if bool(os.getenv('FPAKMAN_UPDATE_NOTIFICATION', 1)):
+        os.system("notify-send {} '{}'".format("-i {}".format(icon_path) if icon_path else '', msg))